Alter and reorganize directory structure
Now uses a separate Make.json to specify the build,
even though its structure and features are still
tightly bound to make.php.

Code not written by me, or code that could be re-used
is now stored under lib/, the parts that make up the
userscript are stored under uso/, the script in itself
is made up of /init.js and /main.js. The build result
is stored in bin/ (even though it's not binary).